Añadido T03 - API Rest

+++ ejemplos/T03_api_rest/movies/src/main/java/es/cesguiro/movies/http_response/Response.java
@@ -26,7 +26,7 @@ public class Response {
     private void buildPaginationMetaData(int totalRecords, int pageSize, int page) {
         this.page = page;
         this.pageSize = pageSize;
-        int totalPages = (int) (Math.ceil(totalRecords / pageSize));
+        int totalPages = (int) (Math.ceil((double) totalRecords / pageSize));
         this.totalPages = totalPages;
 
         if(page > 1 && totalPages > 1)
